[16.0][IMP] purchase_kmitl: merge order/contract/invoice_plan into core module
Consolidate purchase_order_kmitl, purchase_contract_kmitl and
purchase_invoice_plan_kmitl into the core purchase_kmitl module, which now
depends on purchase_manual_delivery.

- Merge all models (purchase.order overrides, purchase.contract.type,
  purchase.invoice.plan, ir.attachment), wizards, data (contract types,
  exception rules, fines cron, decimal precision) and security into
  purchase_kmitl, preserving xmlids.
- Consolidate duplicate purchase.order view inherits: one form, one tree
  (deduplicating work_end_display), one search; relocate the remaining
  base-specific inherits and the contract-type / invoice-plan views.
- Merge the four Thai translation catalogs into a single th.po and rewrite
  references to the removed module xmlids.
- Update all reverse dependencies to depend on purchase_kmitl and fix the
  external view reference in purchase_order_disbursement.
