--- purchase_contract_kmitl/views/purchase_order_views.xml
@@ -1,148 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="view_purchase_order_form_contract" model="ir.ui.view">
-        <field name="name">view.purchase.order.form (Contract KMITL)</field>
-        <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase.purchase_order_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='date_order']/../.." position="after">
-                <group
-                    string="Contract Details"
-                    name="contract_info_section"
-                >
-                    <group>
-                        <field name="contract_name" />
-                        <field
-                            name="contract_type_id"
-                            required="1"
-                            options="{'no_open': True}"
-                        />
-                        <field name="contract_number" />
-                        <field name="date_order_date" required="1" />
-                        <label for="contract_period_days" />
-                        <div name="contract_period_days_div" class="o_row">
-                            <field name="contract_period_days" required="1" />
-                            <span>วัน</span>
-                        </div>
-                        <label for="work_start" string="Date Work" />
-                        <div name="work_start_div" class="o_row">
-                            <field name="work_start" required="1" />
-                            <i
-                                class="fa fa-long-arrow-right mx-2 oe_edit_only"
-                                aria-label="Arrow icon"
-                                title="Arrow"
-                            />
-                            <field name="work_end" readonly="1" />
-                        </div>
-                    </group>
-                </group>
-                <group string="Fines Details" name="fines_supervision_section">
-                    <group>
-                        <field name="fines_late" invisible="1" />
-                        <field name="late_days" invisible="1" />
-                        <field name="fines_rate" widget="float" />
-                        <field name="supervision_cost" widget="float" />
-                    </group>
-                </group>
-            </xpath>
-            <xpath expr="//sheet/notebook" position="inside">
-                <page string="Fines" name="late_fines_page">
-                    <group>
-                        <field name="fines_late" widget="float" readonly="1" />
-                        <field name="late_days" readonly="1" />
-                    </group>
-                </page>
-            </xpath>
-
-            <!-- invisible -->
-            <xpath expr="//field[@name='date_planned']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-            <xpath expr="//field[@name='date_order']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-            <xpath expr="//label[@for='date_planned']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-            <xpath expr="//div[@name='reminder']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-            <xpath expr="//div[@name='date_planned_div']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-            <xpath expr="//field[@name='ignore_exception']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-            <xpath
-                expr="//button[@name='%(purchase_stock.action_purchase_vendor_delay_report)d']"
-                position="attributes"
-            >
-                <attribute name="invisible">1</attribute>
-            </xpath>
-        </field>
-    </record>
-
-    <record id="purchase_order_view_search_kmitl" model="ir.ui.view">
-        <field name="name">purchase.order.search.kmitl (Contract KMITL)</field>
-        <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase.purchase_order_view_search" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='name']" position="after">
-                <field name="contract_number" />
-            </xpath>
-            <xpath expr="//search" position="inside">
-                <filter
-                    string="Work delayed"
-                    name="work_delayed"
-                    domain="[('state', 'in', ['purchase', 'done']), ('late_days', '&gt;', 0)]"
-                />
-            </xpath>
-            <xpath expr="//group" position="inside">
-                <filter
-                    string="Contract Type"
-                    name="contract_type_id"
-                    context="{'group_by': 'contract_type_id'}"
-                />
-            </xpath>
-            <xpath expr="//separator[1]" position="after">
-                <filter
-                    string="Contract period 0 - 15 days"
-                    name="period_0_15"
-                    domain="[('contract_period_days', '&gt;=', 0), ('contract_period_days', '&lt;=', 15)]"
-                />
-                <filter
-                    string="Contract period 0 - 30 days"
-                    name="period_0_30"
-                    domain="[('contract_period_days', '&gt;=', 0), ('contract_period_days', '&lt;=', 30)]"
-                />
-                <filter
-                    string="Contract period 0 - 60 days"
-                    name="period_0_60"
-                    domain="[('contract_period_days', '&gt;=', 0), ('contract_period_days', '&lt;=', 60)]"
-                />
-                <filter
-                    string="Contract period 60+ days"
-                    name="period_60"
-                    domain="[('contract_period_days', '&gt;', 60)]"
-                />
-            </xpath>
-        </field>
-    </record>
-
-    <record id="purchase_order_tree_kmitl" model="ir.ui.view">
-        <field name="name">purchase.order.tree (Contract KMITL)</field>
-        <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase.purchase_order_view_tree" />
-        <field name="arch" type="xml">
-            <xpath expr="//tree" position="inside">
-                <field name="late_days" optional="hide" />
-            </xpath>
-            <xpath expr="//field[@name='date_planned']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-            <xpath expr="//field[@name='amount_total']" position="after">
-                <field name="work_end_display"/>
-            </xpath>
-        </field>
-    </record>
-</odoo>
