--- purchase_invoice_plan_kmitl/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import models, wizard
