--- purchase_invoice_plan_kmitl/__manifest__.py
@@ -1,18 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    "name": "Purchase Invoice Plan KMITL",
-    "version": "16.0.1.0.0",
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "category": "KMITL",
-    "depends": ["purchase_invoice_plan", "purchase_contract_kmitl"],
-    "data": [
-        "data/decimal.xml",
-        "views/purchase_invoice_plan_views.xml",
-        "wizard/purchase_create_invoice_plan_views.xml",
-    ],
-    "application": False,
-    "installable": True,
-    "auto_install": False,
-    "license": "LGPL-3",
-}
