--- purchase_invoice_plan_kmitl/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import purchase_create_invoice_plan
