+++ purchase_kmitl/__manifest__.py
@@ -1,16 +1,39 @@
 # -*- coding: utf-8 -*-
 {
-    'name': 'Purchase KMITL',
-    'version': '16.0.1.0.0',
-    'summary': """ Purchase KMITL Summary """,
+    "name": "Purchase KMITL",
+    "version": "16.0.1.0.0",
+    "summary": """ Purchase KMITL Summary """,
     "category": "KMITL",
     "author": "Aginix Technologies",
     "website": "https://github.com/aginix/kmitl",
-    'depends': ['purchase'],
-    'data': [
-        "views/purchase.xml"
+    "depends": [
+        "purchase_manual_delivery",
+        "purchase_stock",
+        "purchase_requisition",
+        "purchase_no_rfq",
+        "purchase_invoice_plan",
+        "purchase_exception",
+        "purchase_operating_unit",
+        "purchase_request_department",
+        "purchase_order_link_purchase_request",
+        "account_fiscal_year",
+        "hr",
+        "hr_department_short_name",
     ],
-    'installable': True,
-    'auto_install': False,
-    'license': 'LGPL-3',
+    "data": [
+        "security/ir.model.access.csv",
+        "data/decimal.xml",
+        "data/purchase_contract_type.xml",
+        "data/purchase_exception.xml",
+        "data/cron.xml",
+        "views/purchase.xml",
+        "views/purchase_contract_type_views.xml",
+        "views/purchase_order_views.xml",
+        "views/purchase_order_views_invisible.xml",
+        "views/purchase_invoice_plan_views.xml",
+        "wizards/purchase_create_invoice_plan_views.xml",
+    ],
+    "installable": True,
+    "auto_install": False,
+    "license": "LGPL-3",
 }
