+++ purchase_kmitl/data/decimal.xml
