+++ purchase_kmitl/models/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 from . import purchase_order
 from . import purchase_contract_type
+from . import purchase_invoice_plan
+from . import ir_attachment
