+++ purchase_kmitl/models/ir_attachment.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+from odoo import fields, models
+
+
+class IrAttachment(models.Model):
+    _inherit = "ir.attachment"
+
+    sequence = fields.Integer(string="Sequence")
