+++ purchase_kmitl/models/purchase_invoice_plan.py
@@ -2,14 +2,13 @@
 import logging
 from datetime import timedelta
 
-from odoo import _, api, fields, models
-from odoo.exceptions import UserError, ValidationError
+from odoo import api, fields, models
 
 _logger = logging.getLogger(__name__)
 
 
 class PurchaseInvoicePlan(models.Model):
-    _inherit = 'purchase.invoice.plan'
+    _inherit = "purchase.invoice.plan"
 
     duration_days = fields.Integer(string="Duration (Days)")
     plan_date = fields.Date(required=False)
@@ -17,7 +16,9 @@ class PurchaseInvoicePlan(models.Model):
     @api.onchange("duration_days")
     def _onchange_duration_days(self):
         if self.duration_days and self.purchase_id.work_start:
-            self.plan_date = self.purchase_id.work_start + timedelta(days=self.duration_days)
+            self.plan_date = self.purchase_id.work_start + timedelta(
+                days=self.duration_days
+            )
 
     @api.depends("percent")
     def _compute_amount(self):
