+++ purchase_kmitl/wizards/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import purchase_request_line_make_purchase_order
+from . import purchase_create_invoice_plan
