+++ purchase_kmitl/wizards/purchase_create_invoice_plan.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from odoo import api, fields, models
