+++ purchase_kmitl/wizards/purchase_request_line_make_purchase_order.py
@@ -1,18 +1,16 @@
 # -*- coding: utf-8 -*-
-import logging
-
-from odoo import models, fields, api, _
-
-_logger = logging.getLogger(__name__)
+from odoo import api, models
 
 
 class PurchaseRequestLineMakePurchaseOrder(models.TransientModel):
-    _inherit = 'purchase.request.line.make.purchase.order'
+    _inherit = "purchase.request.line.make.purchase.order"
 
     @api.model
     def _prepare_purchase_order(self, picking_type, group_id, company, origin):
         vals = super()._prepare_purchase_order(picking_type, group_id, company, origin)
-        vals.update({
-            "department_id": self.item_ids.request_id.department_id.id,
-        })
+        vals.update(
+            {
+                "department_id": self.item_ids.request_id.department_id.id,
+            }
+        )
         return vals
