--- purchase_order_kmitl/models/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import purchase_order
-from . import ir_attachment
