--- purchase_order_kmitl/models/ir_attachment.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-import logging
-
-from odoo import models, fields, api, _
-from odoo.exceptions import UserError, ValidationError
-
-_logger = logging.getLogger(__name__)
-
-
-class IrAttachment(models.Model):
-    _inherit = 'ir.attachment'
-
-    sequence = fields.Integer(string="Sequence")
