+++ purchase_order_leadtime/__manifest__.py
@@ -6,7 +6,7 @@
     "category": "KMITL",
     "author": "Aginix Technologies",
     "website": "https://github.com/aginix/kmitl",
-    'depends': ['base_state_leadtime', 'purchase_order_kmitl'],
+    'depends': ['base_state_leadtime', 'purchase_kmitl'],
     'application': False,
     'installable': True,
     'auto_install': False,
